refactor: move simple-inventory plugin to Node app
