+++ imports/node-app/plugins/simple-inventory/i18n/en.json
