+++ imports/node-app/plugins/simple-inventory/i18n/index.js
