+++ imports/node-app/plugins/simple-inventory/index.js
@@ -1,11 +1,11 @@
-import i18n from "./i18n";
-import mutations from "./mutations";
-import queries from "./queries";
-import resolvers from "./resolvers";
-import schemas from "./schemas";
-import startup from "./startup";
-import inventoryForProductConfigurations from "./utils/inventoryForProductConfigurations";
-import createDataLoaders from "./utils/createDataLoaders";
+import i18n from "./i18n/index.js";
+import mutations from "./mutations/index.js";
+import queries from "./queries/index.js";
+import resolvers from "./resolvers/index.js";
+import schemas from "./schemas/index.js";
+import startup from "./startup.js";
+import inventoryForProductConfigurations from "./utils/inventoryForProductConfigurations.js";
+import createDataLoaders from "./utils/createDataLoaders.js";
 
 /**
  * @summary Import and call this function to add this plugin to your API.
