+++ imports/node-app/plugins/simple-inventory/mutations/index.js
@@ -1,0 +1,9 @@
+import recalculateReservedSimpleInventory from "./recalculateReservedSimpleInventory.js";
+import updateSimpleInventory from "./updateSimpleInventory.js";
+import updateSimpleInventoryBulk from "./updateSimpleInventoryBulk.js";
+
+export default {
+  recalculateReservedSimpleInventory,
+  updateSimpleInventoryBulk,
+  updateSimpleInventory
+};
