+++ imports/node-app/plugins/simple-inventory/mutations/updateSimpleInventory.js
@@ -1,15 +1,10 @@
-import SimpleSchema from "simpl-schema";
 import ReactionError from "@reactioncommerce/reaction-error";
-import { ProductConfigurationSchema, SimpleInventoryCollectionSchema } from "../simpleSchemas";
-import getReservedQuantity from "../utils/getReservedQuantity";
-
-const inputSchema = new SimpleSchema({
-  productConfiguration: ProductConfigurationSchema,
-  shopId: String
-});
+import { inputSchema } from "../utils/defaults.js";
+import getModifier from "../utils/getMongoUpdateModifier.js";
 
 /**
- * @summary Force recalculation of the system-managed `inventoryReserved` field based on current order statuses.
+ * @summary Updates SimpleInventory data for a product configuration. Pass only
+ *   those arguments you want to update.
  * @param {Object} context App context
  * @param {Object} input Input
  * @param {Object} input.productConfiguration Product configuration object
@@ -19,14 +14,18 @@ const inputSchema = new SimpleSchema({
  * @param {Boolean} input.isEnabled Whether the SimpleInventory plugin should manage inventory for this product configuration
  * @param {Number} input.lowInventoryWarningThreshold The "low quantity" flag will be applied to this product configuration
  *   when the available quantity is at or below this threshold.
- * @returns {Object} Updated inventory values
+ * @param {Object} [options] Other options
+ * @param {Boolean} [options.returnUpdatedDoc=true] Set to `false` as a performance optimization
+ *   if you don't need the updated document returned.
+ * @returns {Object|null} Updated inventory values, or `null` if `returnUpdatedDoc` is `false`
  */
-export default async function recalculateReservedSimpleInventory(context, input) {
+export default async function updateSimpleInventory(context, input, options = {}) {
   inputSchema.validate(input);
 
   const { appEvents, collections, isInternalCall, userHasPermission, userId } = context;
   const { Products, SimpleInventory } = collections;
   const { productConfiguration, shopId } = input;
+  const { returnUpdatedDoc = true } = options;
 
   if (!isInternalCall) {
     // Verify that the product exists. For internal calls, we assume we can skip this
@@ -49,31 +48,39 @@ export default async function recalculateReservedSimpleInventory(context, input)
     }
   }
 
-  const inventoryReserved = await getReservedQuantity(context, productConfiguration);
-
-  const modifier = {
-    $set: {
-      inventoryReserved,
-      updatedAt: new Date()
-    }
-  };
-
-  SimpleInventoryCollectionSchema.validate(modifier, { modifier: true });
-
-  const { value: updatedDoc } = await SimpleInventory.findOneAndUpdate(
+  const modifier = getModifier(input);
+  const { upsertedCount } = await SimpleInventory.updateOne(
     {
       "productConfiguration.productVariantId": productConfiguration.productVariantId,
       shopId
     },
     modifier,
     {
-      returnOriginal: false
+      upsert: true
     }
   );
 
-  if (!updatedDoc) throw new ReactionError("not-tracked", "Inventory not tracked for this product");
+  // If we inserted, set the "reserved" quantity to what it should be. We could have
+  // put this in the $setOnInsert but then we'd have to do the Orders lookup for
+  // calculating reserved every time, even when only an update happens. It's better
+  // to wait until here when we know whether we inserted.
+  if (upsertedCount === 1) {
+    await context.mutations.recalculateReservedSimpleInventory(context, {
+      productConfiguration,
+      shopId
+    });
+  } else {
+    // Only emit event if not upserting, as `recalculateReservedSimpleInventory` already emits it.
+    await appEvents.emit("afterInventoryUpdate", { productConfiguration, updatedBy: userId });
+  }
 
-  await appEvents.emit("afterInventoryUpdate", { productConfiguration, updatedBy: userId });
+  let updatedDoc = null;
+  if (returnUpdatedDoc) {
+    updatedDoc = await SimpleInventory.findOne({
+      "productConfiguration.productVariantId": productConfiguration.productVariantId,
+      shopId
+    });
+  }
 
   return updatedDoc;
 }
