+++ imports/node-app/plugins/simple-inventory/queries/index.js
@@ -1,0 +1,5 @@
+import simpleInventory from "./simpleInventory.js";
+
+export default {
+  simpleInventory
+};
