+++ imports/node-app/plugins/simple-inventory/queries/simpleInventory.js
@@ -1,6 +1,6 @@
 import SimpleSchema from "simpl-schema";
 import ReactionError from "@reactioncommerce/reaction-error";
-import { ProductConfigurationSchema } from "../simpleSchemas";
+import { ProductConfigurationSchema } from "../simpleSchemas.js";
 
 const inputSchema = new SimpleSchema({
   productConfiguration: ProductConfigurationSchema,
