+++ imports/node-app/plugins/simple-inventory/resolvers/Mutation/index.js
@@ -1,0 +1,7 @@
+import recalculateReservedSimpleInventory from "./recalculateReservedSimpleInventory.js";
+import updateSimpleInventory from "./updateSimpleInventory.js";
+
+export default {
+  recalculateReservedSimpleInventory,
+  updateSimpleInventory
+};
