+++ imports/node-app/plugins/simple-inventory/resolvers/Query/index.js
@@ -1,0 +1,5 @@
+import simpleInventory from "./simpleInventory.js";
+
+export default {
+  simpleInventory
+};
