+++ imports/node-app/plugins/simple-inventory/resolvers/index.js
@@ -1,0 +1,7 @@
+import Mutation from "./Mutation/index.js";
+import Query from "./Query/index.js";
+
+export default {
+  Mutation,
+  Query
+};
