+++ imports/node-app/plugins/simple-inventory/schemas/index.js
