+++ imports/node-app/plugins/simple-inventory/schemas/schema.graphql
