+++ imports/node-app/plugins/simple-inventory/simpleSchemas.js
