+++ imports/node-app/plugins/simple-inventory/utils/inventoryForProductConfigurations.js
