+++ imports/node-app/plugins/simple-inventory/utils/inventoryForProductConfigurations.test.js
@@ -1,4 +1,4 @@
-import inventoryForProductConfigurations from "./inventoryForProductConfigurations";
+import inventoryForProductConfigurations from "./inventoryForProductConfigurations.js";
 
 const context = {
   dataLoaders: {
