--- imports/plugins/included/simple-inventory/server/no-meteor/mutations/index.js
@@ -1,9 +1,0 @@
-import recalculateReservedSimpleInventory from "./recalculateReservedSimpleInventory";
-import updateSimpleInventory from "./updateSimpleInventory";
-import updateSimpleInventoryBulk from "./updateSimpleInventoryBulk";
-
-export default {
-  recalculateReservedSimpleInventory,
-  updateSimpleInventoryBulk,
-  updateSimpleInventory
-};
