--- imports/plugins/included/simple-inventory/server/no-meteor/resolvers/Mutation/index.js
@@ -1,7 +1,0 @@
-import recalculateReservedSimpleInventory from "./recalculateReservedSimpleInventory";
-import updateSimpleInventory from "./updateSimpleInventory";
-
-export default {
-  recalculateReservedSimpleInventory,
-  updateSimpleInventory
-};
