--- imports/plugins/included/simple-inventory/server/no-meteor/resolvers/Query/index.js
@@ -1,5 +1,0 @@
-import simpleInventory from "./simpleInventory";
-
-export default {
-  simpleInventory
-};
