--- imports/plugins/included/simple-inventory/server/no-meteor/resolvers/index.js
@@ -1,7 +1,0 @@
-import Mutation from "./Mutation";
-import Query from "./Query";
-
-export default {
-  Mutation,
-  Query
-};
